Refactor programs for readability
The programs were unreadable and hard to change. By changing the names
and introducing bus variables, the programs are more manageable to
switch the bus or device. Additionally, improve the auto-complete of the
directories generated--each program will now have a sub dir under the
device + bus.d directory for faster navigation to output.